requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
google-generativeai==0.8.3
pillow==10.4.0
httpx==0.27.0
